Smoke test for end to end sample load/print file generation
